+++ src/merging/merging_grid.jl
@@ -176,7 +176,7 @@ in each velocity direction
 GridN2Merge(N::Int, extent_multiplier::Float64) = GridN2Merge(N, N, N, extent_multiplier)
 
 """
-    compute_velocity_extent!(merging_grid, cell, species, species_data, phys_props)
+    compute_velocity_extent!(merging_grid, cell, species, species_data, phys_props::PhysProps)
 
 Compute extent of velocity grid based on temperature in the cell.
 
@@ -187,12 +187,35 @@ Compute extent of velocity grid based on temperature in the cell.
 * `species_data`: the array of `Species` data
 * `phys_props`: the `PhysProps` instance containing the computed temperature
 """
-function compute_velocity_extent!(merging_grid, cell, species, species_data, phys_props)
+function compute_velocity_extent!(merging_grid, cell, species, species_data, phys_props::PhysProps)
     @inbounds dv = merging_grid.extent_multiplier .* sqrt.(2 * phys_props.T[cell, species] * k_B / species_data[species].mass)
     @inbounds merging_grid.extent_v_lower = phys_props.v[:, cell, species] .- dv
     @inbounds merging_grid.extent_v_upper = phys_props.v[:, cell, species] .+ dv
     @inbounds merging_grid.extent_v_mid = phys_props.v[:, cell, species]
     @inbounds merging_grid.Δv = SVector{3}(2 * dv[1] / merging_grid.Nx, 2 * dv[2] / merging_grid.Ny, 2 * dv[3] / merging_grid.Nz)
+    merging_grid.Δv_inv = 1.0 ./ merging_grid.Δv
+end
+
+"""
+    compute_velocity_extent!(merging_grid, vx_extent, vy_extent, vz_extent)
+
+Compute extent of velocity grid based on explicitly set extents.
+
+# Positional arguments:
+* `merging_grid`: the grid merging (`GridN2Merge`) instance for which to compute the extent
+* `vx_extent`: lower and upper bounds of the grid extent in the x velocity direction
+* `vy_extent`: lower and upper bounds of the grid extent in the y velocity direction
+* `vz_extent`: lower and upper bounds of the grid extent in the z velocity direction
+"""
+function compute_velocity_extent!(merging_grid, vx_extent, vy_extent, vz_extent)
+    @inbounds merging_grid.extent_v_lower = SVector(vx_extent[1], vy_extent[1], vz_extent[1])
+    @inbounds merging_grid.extent_v_upper = SVector(vx_extent[2], vy_extent[2], vz_extent[2])
+    @inbounds merging_grid.extent_v_mid = SVector(0.5 * (vx_extent[1] + vx_extent[2]),
+                                                  0.5 * (vy_extent[1] + vy_extent[2]),
+                                                  0.5 * (vz_extent[1] + vz_extent[2]))
+    @inbounds merging_grid.Δv = SVector{3}((vx_extent[2] - vx_extent[1]) / merging_grid.Nx,
+                                           (vy_extent[2] - vy_extent[1]) / merging_grid.Ny,
+                                           (vz_extent[2] - vz_extent[1]) / merging_grid.Nz)
     merging_grid.Δv_inv = 1.0 ./ merging_grid.Δv
 end
 
@@ -548,7 +571,7 @@ function compute_new_particles!(rng, merging_grid::GridN2Merge, particles, pia, 
 end
 
 """
-    merge_grid_based!(rng, merging_grid, particles, pia, cell, species, species_data, phys_props)
+    merge_grid_based!(rng, merging_grid, particles, pia, cell, species, species_data, phys_props::PhysProps)
 
 Merge particles using a velocity grid-based merging approach. A Cartesian grid in velocity
 space is used to group particles together (particles outside of the grid are group by velocity
@@ -571,7 +594,7 @@ in the physical grid cell being considered, as stored in the `phys_props` parame
     [Comput. Phys. Comm., 2015](https://doi.org/10.1016/j.cpc.2015.01.020).
 * G. Oblapenko, D. Goldstein, P. Varghese, C. Moore, A velocity space hybridization-based Boltzmann equation solver. [J. Comput. Phys, 2020](https://doi.org/10.1016/j.jcp.2020.109302).
 """
-function merge_grid_based!(rng, merging_grid, particles, pia, cell, species, species_data, phys_props)
+function merge_grid_based!(rng, merging_grid, particles, pia, cell, species, species_data, phys_props::PhysProps)
     # 0-D, no grid, particles in single cell
     compute_velocity_extent!(merging_grid, cell, species, species_data, phys_props)
     compute_grid!(merging_grid, particles, pia, cell, species)
@@ -580,7 +603,39 @@ end
 
 
 """
-    merge_grid_based!(rng, merging_grid, particles, pia, cell, species, species_data, phys_props, grid::Grid1DUniform)
+    merge_grid_based!(rng, merging_grid, particles, pia, cell, species, species_data, vx_extent, vy_extent, vz_extent)
+
+Merge particles using a velocity grid-based merging approach. A Cartesian grid in velocity
+space is used to group particles together (particles outside of the grid are group by velocity
+octant), and in each cell/octant, particles are merged down to 2 particles.
+The extent of the grid is specified explicitly.
+
+# Positional arguments:
+* `rng`: the random number generator instance
+* `merging_grid`: the grid merging (`GridN2Merge`) instance defining the velocity space grid
+* `particles`: the `ParticleVector` instance of the particles to be merged
+* `pia`: the `ParticleIndexerArray` instance
+* `cell`: the cell index
+* `species`: the species index
+* `species_data`: the array of `Species` data
+* `vx_extent`: lower and upper bounds of the grid extent in the x velocity direction
+* `vy_extent`: lower and upper bounds of the grid extent in the y velocity direction
+* `vz_extent`: lower and upper bounds of the grid extent in the z velocity direction
+
+# References
+* M. Vranic, T. Grismayer, J.L. Martins, R.A. Fonseca, L.O. Silva, Particle merging algorithm for PIC codes.
+    [Comput. Phys. Comm., 2015](https://doi.org/10.1016/j.cpc.2015.01.020).
+* G. Oblapenko, D. Goldstein, P. Varghese, C. Moore, A velocity space hybridization-based Boltzmann equation solver. [J. Comput. Phys, 2020](https://doi.org/10.1016/j.jcp.2020.109302).
+"""
+function merge_grid_based!(rng, merging_grid, particles, pia, cell, species, species_data, vx_extent, vy_extent, vz_extent)
+    # 0-D, no grid, particles in single cell
+    compute_velocity_extent!(merging_grid, vx_extent, vy_extent, vz_extent)
+    compute_grid!(merging_grid, particles, pia, cell, species)
+    compute_new_particles!(rng, merging_grid, particles, pia, cell, species)
+end
+
+"""
+    merge_grid_based!(rng, merging_grid, particles, pia, cell, species, species_data, phys_props::PhysProps, grid::Grid1DUniform)
 
 Merge particles using a velocity grid-based merging approach. A Cartesian grid in velocity
 space is used to group particles together (particles outside of the grid are group by velocity
@@ -605,9 +660,43 @@ If particle positions end up outside of the simulation domain, the particles are
     [Comput. Phys. Comm., 2015](https://doi.org/10.1016/j.cpc.2015.01.020).
 * G. Oblapenko, D. Goldstein, P. Varghese, C. Moore, A velocity space hybridization-based Boltzmann equation solver. [J. Comput. Phys, 2020](https://doi.org/10.1016/j.jcp.2020.109302).
 """
-function merge_grid_based!(rng, merging_grid, particles, pia, cell, species, species_data, phys_props, grid::Grid1DUniform)
+function merge_grid_based!(rng, merging_grid, particles, pia, cell, species, species_data, phys_props::PhysProps, grid::Grid1DUniform)
     # 0-D, no grid, particles in single cell
     compute_velocity_extent!(merging_grid, cell, species, species_data, phys_props)
+    compute_grid!(merging_grid, particles, pia, cell, species)
+    compute_new_particles!(rng, merging_grid, particles, pia, cell, species, grid)
+end
+
+"""
+    merge_grid_based!(rng, merging_grid, particles, pia, cell, species, species_data, vx_extent, vy_extent, vz_extent, grid::Grid1DUniform)
+
+Merge particles using a velocity grid-based merging approach. A Cartesian grid in velocity
+space is used to group particles together (particles outside of the grid are group by velocity
+octant), and in each cell/octant, particles are merged down to 2 particles.
+The extent of the grid is specified explicitly.
+If particle positions end up outside of the simulation domain, the particles are placed back into the domain.
+
+# Positional arguments:
+* `rng`: the random number generator instance
+* `merging_grid`: the grid merging (`GridN2Merge`) instance defining the velocity space grid
+* `particles`: the `ParticleVector` instance of the particles to be merged
+* `pia`: the `ParticleIndexerArray` instance
+* `cell`: the cell index
+* `species`: the species index
+* `species_data`: the array of `Species` data
+* `vx_extent`: lower and upper bounds of the grid extent in the x velocity direction
+* `vy_extent`: lower and upper bounds of the grid extent in the y velocity direction
+* `vz_extent`: lower and upper bounds of the grid extent in the z velocity direction
+* `grid`: the `Grid1DUniform` grid
+
+# References
+* M. Vranic, T. Grismayer, J.L. Martins, R.A. Fonseca, L.O. Silva, Particle merging algorithm for PIC codes.
+    [Comput. Phys. Comm., 2015](https://doi.org/10.1016/j.cpc.2015.01.020).
+* G. Oblapenko, D. Goldstein, P. Varghese, C. Moore, A velocity space hybridization-based Boltzmann equation solver. [J. Comput. Phys, 2020](https://doi.org/10.1016/j.jcp.2020.109302).
+"""
+function merge_grid_based!(rng, merging_grid, particles, pia, cell, species, species_data, vx_extent, vy_extent, vz_extent, grid::Grid1DUniform)
+    # 0-D, no grid, particles in single cell
+    compute_velocity_extent!(merging_grid, vx_extent, vy_extent, vz_extent)
     compute_grid!(merging_grid, particles, pia, cell, species)
     compute_new_particles!(rng, merging_grid, particles, pia, cell, species, grid)
 end