+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Speed-up of variable weight SWPM collisions
 * Rate-preserving merging now makes use of reference process cross-sections
 * Fix in `simulations/OD/BKW/bkw.jl`
+* Clean-up of some `@inbounds` macros in sampling routines
 
 ## v0.7.5
 * Octree and grid-based merging can now handle bins/cells with only zero-weight particles inside