+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Rate-preserving merging now makes use of reference process cross-sections
 * Fix in `simulations/OD/BKW/bkw.jl`
 * Clean-up of some `@inbounds` macros in sampling routines
+* Option to set merging grid extent explicitly when calling `merge_grid_based!`
+* Improved test coverage
 
 ## v0.7.5
 * Octree and grid-based merging can now handle bins/cells with only zero-weight particles inside